+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Traits
   - JsonSchemaTrait : Providing JSON Schema generation and validation capabilities for classes.
 - Utils
+  - CborSerializer tool : cbor serializer helper
   - JsonSerializer tool : an helper class to serialize objects with temporary options.
 
 ### Changed